+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- DeepL Integration
+
 ### Fixed
 - Error thrown with PHP 8 due to an optional parameter coming before a required one.